Add task for starting Jenkins
The introduction of this task will assist in eliminating both the
restart.yml and wait-for-start.yml tasks. This task handles both
starting the Jenkins service and waiting for it to complete starting
(formerly handled by wait-for-start.yml). In this manner, we can
achieve a more fine-grained control of the Jenkins service status.