Put Jenkins in correct state in task files
Previous to this change, main.yml attempted to manage Jenkins' service
state before including any of the other task files. This got a bit
messy and also led to some cases of unnecessary service restarts.

Now, each task file will place Jenkins in the desired state before
attempting to do any work. This guarantees that no matter what state
the service is in when a task finishes, the next task will execute
without problems. This also makes it possible for playbooks including
this role to be able to use `include_tasks`/`tasks_from` from this
role without needing to worry about putting the Jenkins service in the
correct state beforehand.